requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
crypto_cpp_py==1.4.4
ecdsa==0.18.0
frozenlist==1.4.1
idna==3.6
lark==1.1.9
marshmallow==3.20.2
marshmallow-dataclass==8.6.0
marshmallow-oneofschema==3.1.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
poseidon_py==0.1.4
pycryptodome==3.20.0
python-decouple==3.8
six==1.16.0
starknet-py==0.19.0
sympy==1.11.1
toml==0.10.2
typing-inspect==0.9.0
typing_extensions==4.9.0
yarl==1.9.4